Organizing some actor folder files
