+++ Assets/Code/Unity/Actor/Attributes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ed3c07b4b898eb54ca5db217afb87905
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
