+++ Assets/Code/Unity/Actor/Enemy/EnemySpawner.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -8,6 +9,7 @@ namespace SaloonSlingers.Unity.Actor
     public class EnemySpawner : MonoBehaviour
     {
         public List<Transform> SpawnPoints;
+        public EventHandler AllEnemiesKilled;
 
         [SerializeField]
         private float spawnPerSecond = 4f;
@@ -29,6 +31,7 @@ namespace SaloonSlingers.Unity.Actor
         private bool currentlyVisible = false;
         private GameManager gameManager;
         private readonly IDictionary<string, ActorPool> pools = new Dictionary<string, ActorPool>();
+        private IDictionary<string, int> enemiesLeft;
 
         public GameObject Spawn(string enemyStr) => pools[enemyStr].Get(false);
 
@@ -42,10 +45,16 @@ namespace SaloonSlingers.Unity.Actor
             return go;
         }
 
-        private void RecordDeath(object sender, System.EventArgs e)
+        private void RecordDeath(object sender, EventArgs e)
         {
             var go = (GameObject)sender;
-            gameManager.Saloon.EnemyInventory.RecordDeath(go.name);
+            enemiesLeft[go.name]--;
+            if (enemiesLeft[go.name] == 0)
+            {
+                enemiesLeft.Remove(go.name);
+                if (enemiesLeft.Count == 0)
+                    AllEnemiesKilled?.Invoke(gameObject, EventArgs.Empty);
+            }
 
             var actor = go.GetComponent<IActor>();
             actor.Death -= RecordDeath;
@@ -61,6 +70,8 @@ namespace SaloonSlingers.Unity.Actor
             }
             InvokeRepeating(nameof(SpawnEnemy), 1, spawnPerSecond);
             currentlyVisible = true;
+            gameManager.Saloon.EnemyInventory.Emptied += OnInventoryEmptied;
+            enemiesLeft = new Dictionary<string, int>(gameManager.Saloon.EnemyInventory.Manifest);
         }
 
         private void Update()
@@ -78,9 +89,14 @@ namespace SaloonSlingers.Unity.Actor
             }
         }
 
+        private void OnInventoryEmptied(object sender, System.EventArgs e)
+        {
+            CancelInvoke(nameof(SpawnEnemy));
+        }
+
         private void SpawnEnemy()
         {
-            if (pools.Values.Sum(p => p.CountSpanwed) == maxActiveEnemies || gameManager.Saloon.EnemyInventory.Completed) return;
+            if (pools.Values.Sum(p => p.CountSpanwed) == maxActiveEnemies) return;
 
             var enemyStr = gameManager.Saloon.EnemyInventory.GetRandomEnemy();
             if (enemyStr == null)
@@ -89,13 +105,13 @@ namespace SaloonSlingers.Unity.Actor
                 return;
             }
 
-            int randomSpawnpointIndex = Random.Range(0, SpawnPoints.Count - 1);
+            int randomSpawnpointIndex = UnityEngine.Random.Range(0, SpawnPoints.Count - 1);
             Transform spawnPoint = SpawnPoints[randomSpawnpointIndex];
-            Vector3 positionNoise = new(Random.Range(minPositionNoise, maxPositionNoise),
-                                        Random.Range(minPositionNoise, maxPositionNoise),
+            Vector3 positionNoise = new(UnityEngine.Random.Range(minPositionNoise, maxPositionNoise),
+                                        UnityEngine.Random.Range(minPositionNoise, maxPositionNoise),
                                         0f);
             Vector3 spawnPosition = spawnPoint.position + positionNoise;
-            Quaternion rotation = Quaternion.Euler(0, Random.Range(minRotationNoise, maxRotationNoise), 0f);
+            Quaternion rotation = Quaternion.Euler(0, UnityEngine.Random.Range(minRotationNoise, maxRotationNoise), 0f);
             GameObject go = Spawn(spawnPosition, rotation, enemyStr);
             go.SetActive(true);
         }
