+++ Assets/Code/Unity/Actor/Pickup.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a04f2c0ca165eb8428fc639997030d57
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
