+++ Assets/Code/Unity/Actor/Pickup/PickupDropper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 75c853cbbf31c77449535f230812b47d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
