+++ Assets/Code/Unity/Actor/Pickup/PickupOnCollision.cs
@@ -6,9 +6,6 @@ namespace SaloonSlingers.Unity
 {
     public class PickupOnCollision : MonoBehaviour
     {
-        [SerializeField]
-        private LayerMask ignoreMask;
-
         private Pickup pickup;
 
         private void Start()
@@ -18,9 +15,6 @@ namespace SaloonSlingers.Unity
 
         private void OnCollisionEnter(Collision collision)
         {
-            if ((ignoreMask & (1 << collision.gameObject.layer)) != 0)
-                return;
-
             if (collision.gameObject.TryGetComponent<Attributes>(out var attributes) &&
                 attributes.Registry.TryGetValue(AttributeType.Money, out var money))
             {
